moved to own example file
